fix: updated unit price data type to float, to match the data on products table

+++ database/migrations/2025_09_01_022847_create_order_items_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
 
             // Pricing
             $table->integer('quantity');
-            $table->decimal('unit_price', 20, 2)->nullable();
+            $table->float('unit_price')->nullable();
             $table->decimal('total_price', 20, 2)->nullable();
 
             $table->timestamps();
